Tidy up .travis.yml and delete travis_install_gdal.sh
We decided to use the GDAL which is available in the ubuntugis ppa and not built it ourselves.
We therefore are do not need the travis_install_gdal script and the GDALVERSION matrix in .travis.yml
Try out if the travis job log is tidier when i use wget --quiet in combination with travis_wait.
Delete all lines which where used in the travis_install_gdal script.